"Update form-field.scss to use placeholder color variable and add styles for p-dropdown"
This commit updates the `form-field.scss` file to use the `--placeholder-input-color` variable instead of the `--text-desc-color` variable for placeholder text. It also adds styles for the `p-dropdown` component to adjust the padding and border radius. Additionally, it adjusts the position of the clear icon for right-to-left (RTL) text direction and updates the styles for the input icon wrapper in RTL.