Add loader component to display while app is loading
The diff includes changes to app.component.html, app.component.ts, and app.module.ts.

In app.component.html, a new ng-container element has been added that conditionally displays the loader component based on the value of the isLoading property in AppComponent.

In app.component.ts, the isLoading property has been added and initialized to true. After a 1-second delay, it is set to false.

In app.module.ts, LoaderComponent has been imported and added to the declarations array.