"Update clients list, add search functionality and client cards in clients component"
This diff includes updates to the clients list, adding search functionality, and displaying client cards in the clients component. The `ClientsComponent` has been updated to fetch clients from the `clientsService` and display them in a table or as cards. A search functionality has been added to filter the clients list based on the input keyword. The `ClientCardComponent` has been updated to display client information in a card format. The `DashboardNavbarComponent` and `DynamicTableComponent` have also been updated with minor changes. The `roots` configuration has been updated to include a `dashboard` object with the `getClients` endpoint. The `ar.json` and `en.json` files have been updated with a new translation for 'no\_records\_found'. The `variables.scss` file has been updated with a new variable for the placeholder input color.