"Add client functionality to ClientsService and corresponding component"
This diff adds a new method `addClient` to the `ClientsService` which makes a POST request to the API endpoint for adding a client. The corresponding component `AddEditClientComponent` has been updated to use this new method in the `submit` method. Additionally, the `unsubscribe` array has been renamed to `subscriptions` and the `submit` method now extracts form data and calls the `addClient` method with this data. The `handleAddClientSuccess` and `handleAddClientError` methods have been added to handle the response from the API.