"Update ClientsComponent and add router outlet in template"
This commit updates the ClientsComponent by removing the router import and adding the RouterModule to the imports array. Additionally, the template of the ClientsComponent is updated to include a router outlet. This allows for child routes to be loaded within the ClientsComponent. The other changes in the commit, such as the removal of certain methods and properties in the ClientsComponent, are likely unrelated to the main purpose of the commit and can be ignored.