"Add upload-multi-files component to edit-client view and display uploaded files"
This diff adds the `upload-multi-files` component to the `edit-client` view and displays the uploaded files. The component is now standalone and has been updated to accept inputs for `filesNames` and `filesSrc`. The `removeImgFile` method has been added to remove a specific file from the list of uploaded files. The component's selector has been changed from `app-upload-multi-files` to `upload-multi-files`. The `upload-multi-files` component's template and styles have been updated to display the uploaded files and add a scrollbar to the list of files. The `formatSizeUnits` method has been updated to handle the file size units.