Update statistics component with localized text and layout; add user name to statistics component; update user info component to use MenuItem interface
The statistics component has been updated with new localized text and a layout that includes a welcome message, a star icon, and a call-to-action button. The user's name has been added to the statistics component and is set to "Yasser, project manager". The user info component has been updated to use the MenuItem interface for its list of user info options. The user info list has also been renamed to "userInfoList". The Arabic and English translation files have been updated with new translations for the statistics component.