"Update app component, add skeleton component, and update clients service to include getClientById method"
This commit includes changes to the app component, the addition of a skeleton component, and updates to the clients service to include a getClientById method. The app component changes include the removal of a div and the addition of an ng-template, as well as the modification of existing form controls. The skeleton component is added to provide a placeholder for loading data. The clients service is updated to include a method for retrieving a client by its ID.