Update clients-list component HTML to use 'dynamic-table' instead of 'dynamic-table-local-actions'
The diff shows the replacement of the 'dynamic-table-local-actions' component with the 'dynamic-table' component in the clients-list component's HTML template. The new component has some changes in input properties, such as the removal of 'localActionsButtons' and the addition of 'enableFilterDriverStatus'. The 'addHandler' event is also now directly emitted from the 'dynamic-table' component.