Update dynamic table components to use font awesome icons for dropdowns
This commit updates the dynamic table components (dynamic-table, dynamic-table-v2, dynamic-table-local-actions) to use Font Awesome icons for dropdowns. The 'dropdownIcon' property is added to the p-dropdown components and set to 'fas fa-caret-down'. This change affects the appearance of dropdowns in the dynamic table components, making them more consistent with the overall look and feel of the application.